Semantic & style improvement
I have change the name of some HTML tags in order to improve the semantic of the code as a whole (by turning divs into sections and main). Also, I have modified the margins and the flexbox of the "blog-post" class, so everything stays right on the center of the page.